+++ cli.js
@@ -16,12 +16,12 @@ updateNotifier({
 }).notify()
 
 const cli = require('yargs')
-cli.env('AEGIR')
+cli.env('TASEGIR')
   .usage('Usage: $0 <command> [options]')
   .example('$0 build', 'Runs the build command to bundle JS code for the browser.')
   .example('npx $0 build', 'Can be used with `npx` to use a local version')
   .example('$0 test -t webworker -- --browsers Firefox', 'If the command supports `--` can be used to forward options to the underlying tool.')
-  .example('npm test -- -- --browsers Firefox', 'If `npm test` translates to `aegir test -t browser` and you want to forward options you need to use `-- --` instead.')
+  .example('npm test -- -- --browsers Firefox', 'If `npm test` translates to `tasegir test -t browser` and you want to forward options you need to use `-- --` instead.')
   .epilog('Use `$0 <command> --help` to learn more about each command.')
   .commandDir('cmds')
   .demandCommand(1, 'You need at least one command.')
@@ -37,7 +37,7 @@ cli.env('AEGIR')
   .group(['help', 'version', 'debug'], 'Global Options:')
   .wrap(cli.terminalWidth())
   .parserConfiguration({ 'populate--': true })
-  // .strict() // this doesnt allow for cross-env AEGIR_TEST=hello node cli.js test -t node --files 'test/**/*.spec.js' "--bail"
+  // .strict() // this doesnt allow for cross-env TASEGIR_TEST=hello node cli.js test -t node --files 'test/**/*.spec.js' "--bail"
 
 const args = cli.fail((msg, err, yargs) => {
   if (msg) {