+++ cmds/test.js
@@ -11,11 +11,11 @@ module.exports = {
   builder: (yargs) => {
     yargs
       .epilog(EPILOG)
-      .example('aegir test -t webworker', 'Run tests in the browser with Karma inside a webworker.')
-      .example('aegir test -t browser -- --browsers Firefox,Chrome,Safari', 'Tell Karma to run tests in several browsers at the same time.')
-      .example('aegir test -w -t browser -- --browser Chrome', 'Debug tests with watch mode and tell Karma to open Chrome in a non-headless mode.')
+      .example('tasegir test -t webworker', 'Run tests in the browser with Karma inside a webworker.')
+      .example('tasegir test -t browser -- --browsers Firefox,Chrome,Safari', 'Tell Karma to run tests in several browsers at the same time.')
+      .example('tasegir test -w -t browser -- --browser Chrome', 'Debug tests with watch mode and tell Karma to open Chrome in a non-headless mode.')
       .example(
-        'aegir test -t electron-renderer -- --interactive',
+        'tasegir test -t electron-renderer -- --interactive',
         'Debug electron renderer test with a persistent window.'
       )
       .options({