+++ src/update-rc/index.js
@@ -44,7 +44,7 @@ async function updateRc (opts) {
     publish: true,
     ghrelease: true,
     docs: true,
-    ghtoken: opts.ghtoken || process.env.AEGIR_GHTOKEN
+    ghtoken: opts.ghtoken || process.env.TASEGIR_GHTOKEN
   })
 }